Fixes and refactors code
This commit addresses several issues and improves code quality.

A bug in the statistics calculation that caused a crash when no logs were present is fixed by explicitly typing the list in `stats_helper`. It also fixes a typo in a debug message for updating drink logs.

Error handling in the `UrineLog` model is improved to check for null values in `color` and `amount` fields during JSON deserialization.

Unused dependencies, specifically `cupertino_icons`, have been removed to clean up the project. Additionally, the initialization for `StorageService` is now correctly enabled in `main.dart`.